pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "token-copilot"
version = "1.0.2"
description = "Your AI copilot for LLM costs - Modern, plugin-based cost tracking for LangChain, LangGraph, and LlamaIndex with 5 usage patterns"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Sai Kumar Yava", email = "saikumar.geek@gmail.com"},
]
keywords = [
    "llm", "langchain", "langgraph", "llamaindex", "cost-tracking", "budget",
    "ai", "openai", "anthropic", "multi-tenant", "cost-optimization", "analytics",
    "monitoring", "observability", "token-tracking", "budget-enforcement"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "langchain-core>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
]

analytics = [
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
]

streaming = [
    "requests>=2.28.0",
    "kafka-python>=2.0.2",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]

langchain = [
    "langchain-openai>=0.1.0",
]

llamaindex = [
    "llama-index-core>=0.10.0",
    "llama-index-llms-openai>=0.1.0",
]

langgraph = [
    "langgraph>=0.2.0",
]

examples = [
    "langchain-openai>=0.1.0",
    "llama-index-core>=0.10.0",
    "llama-index-llms-openai>=0.1.0",
    "langgraph>=0.2.0",
]

all = [
    "token-copilot[dev,docs,analytics,streaming,langchain,llamaindex,langgraph]",
]

[project.urls]
Homepage = "https://github.com/scionoftech/token-copilot"
Documentation = "https://github.com/scionoftech/token-copilot#readme"
Repository = "https://github.com/scionoftech/token-copilot"
Issues = "https://github.com/scionoftech/token-copilot/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--verbose",
]
# Coverage options (requires pytest-cov):
# Uncomment below if you have pytest-cov installed:
# addopts = [
#     "--cov=src/token_copilot",
#     "--cov-report=html",
#     "--cov-report=term-missing",
#     "--verbose",
# ]

[tool.coverage.run]
source = ["src/token_copilot"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "\\.\\.\\.",
]
